src/test/java/org/alan/algorithm/practice/baekjoon/gold/five/BJ7453Test.java
package org.alan.algorithm.practice.baekjoon.gold.five;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ7453Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        6
                        -45 22 42 -16
                        -41 -27 56 30
                        -36 53 -37 77
                        -36 30 -75 -46
                        26 -38 -10 62
                        -32 -54 -6 45""", "5")
        );
    }
    
    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ7453().solve(input));
    }
}
